requirements.txt
aiodns==2.0.0
aiogram==2.13
aiohttp==3.7.4.post0
aiologger==0.6.1
aiopg==1.2.1
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
emoji==1.2.1
idna==2.10
multidict==5.1.0
psycopg2-binary==2.8.6
pycares==3.2.3
pycparser==2.20
pytz==2021.1
requests==2.25.1
suds-community==0.8.4
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.4
uvloop==0.15.2
yarl==1.6.3
